+++ tests/test_wclass.py
@@ -229,7 +229,7 @@ class ClassTestCase(unittest.TestCase):
             sorted([c1, c2, c3], key=lambda i: i.qclass),
             sorted(Class.list(WIMS_URL, "myself", "toto", "rclass"), key=lambda i: i.qclass)
         )
-
+        
         self.assertListEqual(
             [],
             Class.list(WIMS_URL, "myself", "toto", "unknown_rclass")
@@ -275,7 +275,7 @@ class ClassTestCase(unittest.TestCase):
         s3 = Sheet("Third", "Third one")
         
         c.save(WIMS_URL, "myself", "toto")
-
+        
         self.assertListEqual([], c.listitem(Sheet))
         
         c.additem(s1)